Restructure to keep app in main folder
